Refactor and fix debug CLI test coverage (CORE-69)
Fixes up the test classes in the cli-debug module to split tests for
each debug command into a separate class to make the tests cleaner, and
allows for more test parallelism.  Also adds missing test coverage for
the replay command.
